fix logger writing to hardcoded path

+++ awsume/awsumepy/lib/config_management.py
@@ -29,14 +29,15 @@ operations:
   - set [config_key] [config_value]
 """
 
-
-def load_config() -> dict:
+def migrate_to_xdg_base_directories() -> None:
     if not os.path.exists(str(constants.AWSUME_DIR)):
         os.makedirs(str(constants.AWSUME_DIR))
     if not os.path.exists(str(constants.AWSUME_CONFIG.parent)):
         os.makedirs(str(constants.AWSUME_CONFIG.parent))
     if not os.path.isfile(str(constants.AWSUME_CONFIG)):
         open(str(constants.AWSUME_CONFIG), 'a').close()
+    if not os.path.exists(str(constants.AWSUME_LOG_DIR)):
+        os.makedirs(str(constants.AWSUME_LOG_DIR))
     if constants.IS_USING_XDG_CONFIG_HOME and os.path.isfile(constants.AWSUME_CONFIG_LEGACY_PATH):
         # handle migration to XDG base directory: config file
         move(constants.AWSUME_CONFIG_LEGACY_PATH, constants.AWSUME_CONFIG)
@@ -46,16 +47,21 @@ def load_config() -> dict:
     if constants.IS_USING_XDG_DATA_HOME and os.path.isdir(constants.AWSUME_CACHE_DIR_LEGACY_PATH):
         # handle migration to XDG base directory: data dir
         move(constants.AWSUME_DIR_LEGACY_PATH, constants.AWSUME_DIR)
+    if constants.IS_USING_XDG_STATE_HOME and os.path.isdir(constants.AWSUME_LOG_DIR_LEGACY_PATH):
+        # handle migration to XDG base directory: log dir
+        move(constants.AWSUME_LOG_DIR_LEGACY_PATH, constants.AWSUME_LOG_DIR)
     if (
             constants.IS_USING_XDG_CONFIG_HOME and
             constants.IS_USING_XDG_DATA_HOME and
             constants.IS_USING_XDG_CACHE_HOME and
+            constants.IS_USING_XDG_STATE_HOME and
             os.path.isdir(constants.AWSUME_DIR_LEGACY_PATH)
     ):
         # cleanup legacy ~/.awsume dir
         os.rmdir(constants.AWSUME_DIR_LEGACY_PATH)
 
-
+def load_config() -> dict:
+    migrate_to_xdg_base_directories()
 
     options = None
     try:
